Isolate provider SSE stream parsing
Move the shared SSE data-block parser out of runtime-loader.ts so provider stream adapters share a focused parsing module while keeping their event handling logic unchanged. This continues shrinking the runtime-loader monolith without changing stream semantics.

Constraint: Behavior-preserving refactor only; no stream protocol or event-shape changes
Rejected: Split full stream adapters in this pass | higher-risk because provider-specific state machines are intertwined with usage/tool/reasoning handling
Rejected: Change malformed JSON handling | existing behavior silently drops unparseable data blocks
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep SSE parser tolerant of CRLF and malformed data blocks unless stream compatibility tests are updated deliberately
Tested: VF_DISABLE_LRU_INTERVAL=1 deno test --no-check --allow-all src/provider/runtime-loader.test.ts src/provider/runtime-loader/provider-request-init.test.ts; deno fmt --check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-sse.ts; DENO_NO_PACKAGE_JSON=1 deno lint src/provider/runtime-loader.ts src/provider/runtime-loader/provider-sse.ts; deno check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-sse.ts; deno task typecheck
Not-tested: Full deno task test suite before push; pre-push hook is expected to run it
