+++ src/provider/runtime-loader.ts
@@ -19,6 +19,7 @@ import {
   createGoogleRequestInit,
   createOpenAIRequestInit,
 } from "./runtime-loader/provider-request-init.ts";
+import { parseSseChunk } from "./runtime-loader/provider-sse.ts";
 import type { ProviderKind } from "./runtime-loader/provider-http.ts";
 import { requestJson, requestStream } from "./runtime-loader/provider-http.ts";
 import { readRecord } from "./runtime-loader/provider-records.ts";
@@ -1286,36 +1287,6 @@ function buildAnthropicGenerateResult(payload: unknown): {
     finishReason: normalizeAnthropicFinishReason(record?.stop_reason),
     usage: extractAnthropicUsage(payload),
   };
-}
-
-function parseSseChunk(chunk: string): {
-  events: Array<unknown | "[DONE]">;
-  remainder: string;
-} {
-  const blocks = chunk.split(/\r?\n\r?\n/);
-  const remainder = blocks.pop() ?? "";
-  const events = blocks.flatMap((block) => {
-    const dataLines = block.split(/\r?\n/)
-      .filter((line) => line.startsWith("data:"))
-      .map((line) => line.slice(5).trimStart());
-
-    if (!dataLines.length) {
-      return [];
-    }
-
-    const payload = dataLines.join("\n").trim();
-    if (payload === "[DONE]") {
-      return ["[DONE]" as const];
-    }
-
-    try {
-      return [JSON.parse(payload) as unknown];
-    } catch {
-      return [];
-    }
-  });
-
-  return { events, remainder };
 }
 
 async function* streamAnthropicCompatibleParts(
