+++ src/provider/runtime-loader/provider-sse.ts
@@ -1,0 +1,29 @@
+export function parseSseChunk(chunk: string): {
+  events: Array<unknown | "[DONE]">;
+  remainder: string;
+} {
+  const blocks = chunk.split(/\r?\n\r?\n/);
+  const remainder = blocks.pop() ?? "";
+  const events = blocks.flatMap((block) => {
+    const dataLines = block.split(/\r?\n/)
+      .filter((line) => line.startsWith("data:"))
+      .map((line) => line.slice(5).trimStart());
+
+    if (!dataLines.length) {
+      return [];
+    }
+
+    const payload = dataLines.join("\n").trim();
+    if (payload === "[DONE]") {
+      return ["[DONE]" as const];
+    }
+
+    try {
+      return [JSON.parse(payload) as unknown];
+    } catch {
+      return [];
+    }
+  });
+
+  return { events, remainder };
+}
